Add unittest for trainer
Tests if trainer can be initialized and if trainer runs for one epoch (including validation)
Also tests if folder structure and logging files are created correclty.

This PR also needed to adapt the test data, to be in the right format. Initally, labels and images were identically. Now labels do represent binary labels.