Add unittest for trainer (#59)
* Add unittest for trainer

Tests if trainer can be initialized and if trainer runs for one epoch (including validation)
Also tests if folder structure and logging files are created correclty.

This PR also needed to adapt the test data, to be in the right format. Initally, labels and images were identically. Now labels do represent binary labels.

* add scipy to requirements

Apparently this is needed to allow for metric calculation

* add scipy to requirements

SciPy is needed for Hausdorff and Surface Distance

* Update setup.cfg

* Extend unittests for dataloader

* Update tests/test_data.py

Co-authored-by: Markus Hinsche <m.hinsche@gmail.com>

* rebuttal

* black

Co-authored-by: Markus Hinsche <m.hinsche@gmail.com>